# spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Harnessix Action API",
    "description": "跨 Agent 框架的副作用安全执行与治理平面。",
    "version": "0.1.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "tags": [
          "系统"
        ],
        "summary": "Health",
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Healthz Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tools": {
      "get": {
        "tags": [
          "工具"
        ],
        "summary": "List Tools",
        "operationId": "list_tools_v1_tools_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions": {
      "post": {
        "tags": [
          "Action"
        ],
        "summary": "Submit Action",
        "operationId": "submit_action_v1_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionSnapshot"
                }
              }
            }
          },
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionSnapshot"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions/{action_id}": {
      "get": {
        "tags": [
          "Action"
        ],
        "summary": "Get Action",
        "operationId": "get_action_v1_actions__action_id__get",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionSnapshot"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions/{action_id}/events": {
      "get": {
        "tags": [
          "Action"
        ],
        "summary": "List Action Events",
        "operationId": "list_action_events_v1_actions__action_id__events_get",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions/{action_id}/approval": {
      "post": {
        "tags": [
          "审批"
        ],
        "summary": "Decide Approval",
        "operationId": "decide_approval_v1_actions__action_id__approval_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionSnapshot"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionSnapshot"
                }
              }
            },
            "description": "Accepted"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions/{action_id}/reconcile": {
      "post": {
        "tags": [
          "对账"
        ],
        "summary": "Reconcile Action",
        "operationId": "reconcile_action_v1_actions__action_id__reconcile_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionSnapshot"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionSnapshot"
                }
              }
            },
            "description": "Accepted"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionContext": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Session Id",
            "description": "上游会话标识"
          },
          "run_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Run Id",
            "description": "上游运行标识"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id",
            "description": "链路追踪标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "run_id"
        ],
        "title": "ActionContext"
      },
      "ActionEvent": {
        "properties": {
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "sequence": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sequence"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "from_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_status": {
            "$ref": "#/components/schemas/ActionStatus"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action_id",
          "sequence",
          "event_type",
          "to_status",
          "created_at"
        ],
        "title": "ActionEvent"
      },
      "ActionFailure": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retriable": {
            "type": "boolean",
            "title": "Retriable",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ActionFailure"
      },
      "ActionRequest": {
        "properties": {
          "spec_version": {
            "type": "string",
            "const": "harnessix.action/v1",
            "title": "Spec Version",
            "description": "Action Contract 版本",
            "default": "harnessix.action/v1"
          },
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id",
            "description": "Action 全局唯一标识"
          },
          "tool": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "pattern": "^[a-zA-Z][a-zA-Z0-9_.-]*$",
            "title": "Tool",
            "description": "运行时注册的工具名称"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments",
            "description": "工具参数"
          },
          "principal": {
            "$ref": "#/components/schemas/Principal"
          },
          "context": {
            "$ref": "#/components/schemas/ActionContext"
          },
          "effect_hint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EffectClass"
              },
              {
                "type": "null"
              }
            ],
            "description": "调用方预期副作用；最终分类由运行时定义"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key",
            "description": "租户范围业务幂等键"
          },
          "secret_refs": {
            "items": {
              "$ref": "#/components/schemas/SecretRef"
            },
            "type": "array",
            "title": "Secret Refs",
            "description": "凭据引用"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "非敏感扩展元数据"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tool",
          "principal",
          "context"
        ],
        "title": "ActionRequest"
      },
      "ActionResult": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ActionStatus"
          },
          "output": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionFailure"
              },
              {
                "type": "null"
              }
            ]
          },
          "receipt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EffectReceipt"
              },
              {
                "type": "null"
              }
            ]
          },
          "attempt": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Attempt",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ActionResult"
      },
      "ActionSnapshot": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/ActionRequest"
          },
          "request_fingerprint": {
            "type": "string",
            "title": "Request Fingerprint"
          },
          "tool": {
            "$ref": "#/components/schemas/ToolDescriptor"
          },
          "status": {
            "$ref": "#/components/schemas/ActionStatus"
          },
          "policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PolicyDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "lease_owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lease Owner"
          },
          "lease_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lease Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request",
          "request_fingerprint",
          "tool",
          "status",
          "created_at",
          "updated_at",
          "version"
        ],
        "title": "ActionSnapshot"
      },
      "ActionStatus": {
        "type": "string",
        "enum": [
          "received",
          "validated",
          "policy_evaluated",
          "denied",
          "pending_approval",
          "ready",
          "leased",
          "running",
          "succeeded",
          "failed",
          "unknown",
          "reconciling",
          "manual_intervention"
        ],
        "title": "ActionStatus"
      },
      "ApprovalDecision": {
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/ApprovalOutcome"
          },
          "actor": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Actor"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "outcome",
          "actor"
        ],
        "title": "ApprovalDecision"
      },
      "ApprovalOutcome": {
        "type": "string",
        "enum": [
          "approved",
          "rejected"
        ],
        "title": "ApprovalOutcome"
      },
      "ApprovalRecord": {
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/ApprovalOutcome"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "request_fingerprint": {
            "type": "string",
            "title": "Request Fingerprint"
          },
          "decided_at": {
            "type": "string",
            "format": "date-time",
            "title": "Decided At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "outcome",
          "actor",
          "request_fingerprint"
        ],
        "title": "ApprovalRecord"
      },
      "EffectClass": {
        "type": "string",
        "enum": [
          "read_only",
          "idempotent_write",
          "non_idempotent_write",
          "destructive"
        ],
        "title": "EffectClass"
      },
      "EffectReceipt": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "response_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Digest"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "resource_type"
        ],
        "title": "EffectReceipt"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "EventListResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/ActionEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "EventListResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PolicyDecision": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/PolicyDecisionKind"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "policy_id",
          "reason"
        ],
        "title": "PolicyDecision"
      },
      "PolicyDecisionKind": {
        "type": "string",
        "enum": [
          "allow",
          "deny",
          "require_approval"
        ],
        "title": "PolicyDecisionKind"
      },
      "Principal": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Tenant Id",
            "description": "租户标识"
          },
          "subject_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Subject Id",
            "description": "请求主体标识"
          },
          "framework": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Framework",
            "description": "上游 Agent 框架"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "description": "主体角色"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tenant_id",
          "subject_id",
          "framework"
        ],
        "title": "Principal"
      },
      "RiskLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "RiskLevel"
      },
      "SecretRef": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Name",
            "description": "凭据引用名称"
          },
          "version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "description": "凭据版本"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SecretRef"
      },
      "ToolDescriptor": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "input_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Schema"
          },
          "effect_class": {
            "$ref": "#/components/schemas/EffectClass"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "requires_idempotency": {
            "type": "boolean",
            "title": "Requires Idempotency"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval"
          },
          "supports_reconciliation": {
            "type": "boolean",
            "title": "Supports Reconciliation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "version",
          "description",
          "input_schema",
          "effect_class",
          "risk_level",
          "requires_idempotency",
          "requires_approval",
          "supports_reconciliation"
        ],
        "title": "ToolDescriptor"
      },
      "ToolListResponse": {
        "properties": {
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ToolDescriptor"
            },
            "type": "array",
            "title": "Tools"
          }
        },
        "type": "object",
        "required": [
          "tools"
        ],
        "title": "ToolListResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}